requirements.txt
playwright
python-dotenv
fastapi
uvicorn[standard]
motor
